refactor: Rename crud/event.py to maintain consistency with the project
